Reorg backend configuration.
Reorg backend configuration to allow for different adapters.
