+++ lib/basket/configuration.rb
@@ -3,5 +3,9 @@ module Basket
     def queue_collection
       @queue_collection ||= Basket::QueueCollection.new
     end
+
+    def backend
+      @backend ||= BackendAdapter::HashBackend
+    end
   end
 end
